vehicle_worker/requirements.txt
ultralytics
opencv-python
redis
aio_pika
Pillow
